+++ Sources/ProjectSpec/SpecValidation.swift
@@ -53,8 +53,8 @@ extension Project {
             }
         }
 
-        for (name, package) in localPackages {
-            if !(basePath + Path(package.path).normalize()).exists {
+        for (name, package) in packages {
+            if case let .local(path) = package, !(basePath + Path(path).normalize()).exists {
                 errors.append(.invalidLocalPackage(name))
             }
         }
@@ -173,7 +173,7 @@ extension Project {
                         }
                     }
                 case .package:
-                    if packages[dependency.reference] == nil, localPackages[dependency.reference] == nil {
+                    if packages[dependency.reference] == nil {
                         errors.append(.invalidSwiftPackage(name: dependency.reference, target: target.name))
                     }
                 default: break