+++ Tests/ProjectSpecTests/ProjectSpecTests.swift
@@ -97,7 +97,7 @@ class ProjectSpecTests: XCTestCase {
                 project.settings = invalidSettings
                 project.configFiles = ["invalidConfig": "invalidConfigFile"]
                 project.fileGroups = ["invalidFileGroup"]
-                project.localPackages = ["invalidLocalPackage" : LocalSwiftPackage(path: "invalidLocalPackage")]
+                project.packages = ["invalidLocalPackage" : .local(path: "invalidLocalPackage")]
                 project.settingGroups = ["settingGroup1": Settings(
                     configSettings: ["invalidSettingGroupConfig": [:]],
                     groups: ["invalidSettingGroupSettingGroup"]
@@ -516,12 +516,11 @@ class ProjectSpecTests: XCTestCase {
                                                                                                                  script: "bar",
                                                                                                                  settingsTarget: "foo")]))],
                                    packages: [
-                                       "Yams": SwiftPackage(
-                                           url: "https://github.com/jpsim/Yams",
-                                           versionRequirement: .upToNextMajorVersion("2.0.0")
+                                       "Yams": .remote(
+                                                url: "https://github.com/jpsim/Yams",
+                                                versionRequirement: .upToNextMajorVersion("2.0.0")
                                        ),
                                    ],
-                                   localPackages: ["../../Package" : LocalSwiftPackage(path: "../../Package")],
                                    options: SpecOptions(minimumXcodeGenVersion: Version(major: 3, minor: 4, patch: 5),
                                                         carthageBuildPath: "carthageBuildPath",
                                                         carthageExecutablePath: "carthageExecutablePath",
@@ -561,7 +560,6 @@ class ProjectSpecTests: XCTestCase {
                 try expect(proj.settingGroups) == restoredProj.settingGroups
                 try expect(proj.targets) == restoredProj.targets
                 try expect(proj.packages) == restoredProj.packages
-                try expect(proj.localPackages) == restoredProj.localPackages
 
                 try expect(proj) == restoredProj
             }